refactor: Move backup configuration functions to application module and update main.rs imports
